VCITA2-7983 move subscriber to sub-module with event handling and metrics tracking
