+++ src/event-bus.module.ts
@@ -2,13 +2,15 @@ import { DynamicModule, Module } from '@nestjs/common';
 import { PublisherModule } from './modules/publisher/publisher.module';
 import { EventBusConfig } from './interfaces/event-bus-config.interface';
 import { EVENT_BUS_CONFIG } from './constants';
+import { SubscriberModule } from './modules/subscriber/subscriber.module';
 
+const modules = [PublisherModule, SubscriberModule];
 /**
  * Module providing event bus functionality for publishing events
  */
 @Module({
-  imports: [PublisherModule],
-  exports: [PublisherModule],
+  imports: modules,
+  exports: modules,
 })
 export class EventBusModule {
   /**
@@ -17,14 +19,14 @@ export class EventBusModule {
   static forRoot(config: EventBusConfig): DynamicModule {
     return {
       module: EventBusModule,
-      imports: [PublisherModule],
+      imports: modules,
       providers: [
         {
           provide: EVENT_BUS_CONFIG,
           useValue: config,
         },
       ],
-      exports: [PublisherModule],
+      exports: modules,
     };
   }
 }
