+++ src/index.ts
@@ -12,6 +12,15 @@ export {
   Event,
   PublishEventOptions,
 } from './interfaces/event.interface';
+export {
+  RetryOptions,
+  SubscribeToOptions,
+  LegacySubscribeToOptions,
+} from './interfaces/subscription.interface';
 
 // Type exports
 export { EventBusConfig } from './interfaces/event-bus-config.interface';
+
+// Decorator exports
+export { SubscribeTo } from './modules/subscriber/decorators/subscribe-to.decorator';
+export { LegacySubscribeTo } from './modules/subscriber/decorators/legacy-subscribe-to.decorator';
