+++ src/modules/subscriber/configuration.ts
